+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 0.11.1 (?)
 
+- Updated `property_mask` of filter to use a new type of mask, called
+  `PropertyMask`, which allows one to distinguish between real properties
+  and pseudo-properties.
+
 - Added `describe_mask` option to RDF encoder.  This determines whether
   statements corresponding to entity descriptors should be included in the
   serialization automatically.  The boolean aliases `describe_aliases`,